+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### **BREAKING** Changes
 
 - Completely change option functions: turn `short_name`, `is_optional`, and `default_value` into optional arguments
+- Change C++ option functions to return a builder
 - Make options return constant pointers
 - Rename `ARGS_MIN_DESC_LENGTH` to `ARGS_MIN_DESCRIPTION_LENGTH`
 - Rename `option_str` and `option_enum_str` to `option_string` and `option_enum_string`